Use correct query aggregator in OauthPlugin
The Oauth "string to sign" includes the query string parameters,
together with the OAuth parameters. Guzzle uses a QueryString instance
to combine the two. However, it was incorrectly using the default
aggregator, instead of the actual aggregator used by the QueryString
that's part of the request.

This change sets the aggregator by copying it from the request's
QueryString.