Fixed documentation.
I removed a line from the doc:

    'timeout'         => 10,

This parameter is ignored as a default because 'timeout' is not one of the keys
returned by getDefaultOptions. Thus, it is ignored by the array_replace call.

It is important that this document be correct; otherwise people may think that
they are setting a timeout when in fact their timeout setting is being ignored.