Obfuscates URI if there is an username and a password present
Lets assume that you include username and password into the URI. Currently, if the exception accrued the error message looks like so:

```
Client error: `POST http://username:secrect@hostname.com/somePath` resulted in a `400 Bad Request` response: ......,  (truncated...)
```

It lead to security issue because password is shown like it is.

This commit will hide the password (replace it with `***`).