Update Quickstart.rst to fix syntax issue
I’m just getting started with Guzzle so I might be wrong about that. I just tried to implement a Quickstart example and I believe the given syntax from the example doesn’t work that way. My example didn’t respond as I expected and that change made it work.

`$headers` has to be a nested array itself otherwise it will be ignored by the request. The example doesn’t read as concise as before but that way it should be correct and work out of the box.