request no encoding if user does not specify acceptable encodings
due to (imo shitty) specs, when the client does not request any encodings, the server is free to pick any encoding it wants. not sending any `Accept-Encoding` headers is equivalent to sending
Accept-Encoding: *
(as * - anything - is the default Accept-encoding value)

IMO we shouldn't overwrite curl's Accept-Encoding header at all and the guzzle implementation is stupid, i'd love to hear the guzzle dev's rationale, 
 but at least setting CURLOPT_ENCODING='' + Accept-Encoding: identity is better than CURLOPT_ENCODING='' + Accept-encoding: *  (* is implicit when no accept-encoding is specified)

Lets say for example that curl is compiled with gzip+deflate, and curl *wants* to send `Accept-Encoding: gzip,deflate`, but because Guzzle sets `Accept-Encoding: *` (implicitly), the server opts to use br, resulting in broken downloads. this wouldn't have happened if guzzle didn't overwrite curl's default accept-encoding, and it wouldn't have happened if guzzle sent `Accept-Encoding: identity`, but it may happen if guzzle sends `Accept-Encoding: *` as it does prior to this patch.